requirements.txt
torch==2.7.1+cu126
datasets==3.6.0
transformers==4.52.4
miditok==3.0.5.post1
symusic==0.5.8
python-dotenv==1.1.0
tqdm==4.67.1
pytest==8.3.5
